+++ react/components/organisms/profile/profile.js
@@ -123,7 +123,8 @@ export class Profile extends Component {
             borderBottomLeftRadius: 0,
             borderBottomRightRadius: 0,
             borderTopLeftRadius: 0,
-            borderTopRightRadius: 0
+            borderTopRightRadius: 0,
+            paddingLeft: 15
         };
     };
 
@@ -163,7 +164,7 @@ export class Profile extends Component {
                         items={this._buttons()}
                         orientation={"vertical"}
                         variant={"flat"}
-                        align={"flex-start"}
+                        align={"left"}
                         toggle={false}
                     />
                 </View>